Modify the code according to the review requirements

+++ a1_driver/include/a1_driver/A1_ros.h
@@ -9,13 +9,19 @@
 #include "rclcpp/rclcpp.hpp"
 #include "a1_driver/A1_wrapper.h"
 #include "a1_comm/A1_comm.h"
+#include "geometry_msgs/msg/twist.hpp"
+#include "a1_msgs/srv/mode.hpp"
+#include "a1_msgs/srv/high_state.hpp"
+#include "a1_msgs/srv/imu.hpp"
+#include "a1_msgs/srv/cartesian.hpp"
+#include "a1_msgs/msg/pose.hpp"
 
 class A1ROS{
 public:
     A1ROS(std::string node_name) : wrapper(A1Wrapper()) {
         this->node_name = node_name;
     };
-    A1ROS(std::string node_name, int sport_mode) : wrapper(A1Wrapper(sport_mode)) {
+    A1ROS(std::string node_name, uint8_t sport_mode) : wrapper(A1Wrapper(sport_mode)) {
         this->node_name = node_name;
     };
     int node_init(int argc, char *argv[]);
